Feat: Make cheat module

+++ src/user/user.service.ts
@@ -270,6 +270,9 @@ export class UserService {
 
   async deleteAllNickname(userUuid?: string) {
     if (userUuid) {
+      if (!(await this.findOne(userUuid))) {
+        throw new NotFoundException('유저를 찾을 수 없습니다.');
+      }
       await this.nicknameRepo.delete({ userUuid: userUuid });
     } else {
       await this.nicknameRepo.delete({});
@@ -278,6 +281,9 @@ export class UserService {
 
   async deleteAllAccount(userUuid?: string) {
     if (userUuid) {
+      if (!(await this.findOne(userUuid))) {
+        throw new NotFoundException('유저를 찾을 수 없습니다.');
+      }
       await this.accountRepo.delete({ userUuid: userUuid });
     } else {
       await this.accountRepo.delete({});
